requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
